requirements.txt
langchain
openai
google-search-results
numexpr
python-dotenv